perf(profiling): stack-v2 adaptive sampling (#13110)
We implement a first iteration on adaptive sampling for stack-v2. The
current approach is based on the comparison between the CPU time used by
the sampling thread vs the CPU time used by the rest of the process. The
latter includes the work of our periodic worker processes. However,
these run with period of O(1s) and are mostly on I/O, so we make the
assumption that their contribution towards CPU overhead can be
neglected. The algorithm that we implement aims at reaching an overhead
that is close to the given default threshold by scaling the sampling
interval according to the measured CPU times.

## Subsequent iterations

In subsequent iterations on the adaptive sampling, we can consider the
following alternative approaches that build up on the one proposed in
this PR:

- High-water mark: track the maximum CPU usage and target a slice of
that
- cgroup CPU limits: detect any CPU usage limits set via cgroup and use
it to compute the target

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [ ] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)