perf(profiling): improve scaling of memory profiler for large heaps
The heap profiler component of memalloc uses a plain array to track
allocations. It performs a linear scan of this array on every free to
see if the freed allocation is tracked. This means the cost of freeing
an allocation increases substantially for programs with large heaps.

Switch to an unordered map for tracking allocations. This should have
roughly constant overhead even for larger heaps, where we track more
allocations.

Rather than find or implement a map in C, use C++. We can wrap a
std::unordered_map in a thin compatibility layer and use it from the
rest of memalloc without having to completely refactor memalloc.

Making this work requires some awkward modifications to setup.py to
build memalloc. Basically, the Extension class can have sources from
multiple languages, but it isn't possible to pass different flags per
language, such as the language standard. Instead, we can use a method
borrowed from a Stack Overflow answer and use the build_clib command,
extended to support compiler arguments. This lets us build a static
library, which we can link from memalloc.

Note that with the way we build the library, it'll be passed to the link
step for other C extensions as well. This should be harmless, though.
The other extensions will only include symbols they actually use, which
in practice will be none. We'll only have trouble if there are
duplicated definitions, but the functions in the map library here are
all "namespaced".

Note also that we might get slightly worse performance due to the extra
cost of map access for very small heaps, compared to using a simple
array.
TODO - demonstrate the difference in performance with this change for
large and small heaps.

Fixes #13307