build: try explicitly setting AR
To get memalloc_map to build in all our setups. Some setups in
multiple_os_tests fail with errors finding `llvm-ar`. This might be due
to how the Python interpreter there was built, which affects what the AR
value is from sysconfig. Just set it to plain `ar`. This is desperate...