fix(iast): avoid excessive filtering of stacktrace locations (#13272)
Current stacktrace filtering code returned no location if all frames are
outside the current working directory or in site-packages. This often
happens in the following scenarios:
- asyncio code: the full stacktrace may be in site-packages, outside app
directory.
- running applications installed in site-packages.

In these cases, vulnerabilities were being dropped. The fix includes:
- First, the first frame that is not a frozen package (or `<template>`
or anything starting with `<`), and not in the stdlib path (but
potentially in the purelib path) is found and saved as a backup frame,
if any. If none is found, no location is returned.
- From there, the first frame that is not a frozen package, not in the
stdlib path, and not in the purelib path is checked. If found, that
location is returned. Otherwise, the backup frame is returned.
- Code to relativize path to the vulnerability location is now extended
to account for the fact that we can return paths within the purelib
path.

In practice, this means that vulnerabilities found in site-packages,
with no other relevant frame, will be reported.

[APPSEC-57414](https://datadoghq.atlassian.net/browse/APPSEC-57414)

(cherry picked from commit 20a96bb791aa98657ba85e77032f670a0ea1b028)