fix(iast): avoid excessive filtering of stacktrace locations (#13272)
Current stacktrace filtering code returned no location if all frames are
outside the current working directory or in site-packages. This often
happens in the following scenarios:
- asyncio code: the full stacktrace may be in site-packages, outside app
directory.
- running applications installed in site-packages.

In these cases, vulnerabilities were being dropped. The fix includes:
- First, the first frame that is not a frozen package (or `<template>`
or anything starting with `<`), and not in the stdlib path (but
potentially in the purelib path) is found and saved as a backup frame,
if any. If none is found, no location is returned.
- From there, the first frame that is not a frozen package, not in the
stdlib path, and not in the purelib path is checked. If found, that
location is returned. Otherwise, the backup frame is returned.
- Code to relativize path to the vulnerability location is now extended
to account for the fact that we can return paths within the purelib
path.

In practice, this means that vulnerabilities found in site-packages,
with no other relevant frame, will be reported.

[APPSEC-57414](https://datadoghq.atlassian.net/browse/APPSEC-57414)

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)


[APPSEC-57414]:
https://datadoghq.atlassian.net/browse/APPSEC-57414?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ