profiling: handle missed untrack for heap profiler
If we fail to acquire the heap lock when untracking an allocation, we
won't remove it. The next time we sample an allocation with the same
address, it'll collide with the existing entry. When this happens, free
the old traceback, which corresponds to a freed allocation, and replace
it with the new one.