fix(llmobs): reuse shared conn (#13339)
Resolves #13336. Credit to @IAL32 and a cherry-pick from #13338.

When we made the jump from using the shared HTTPWriter to our own
BaseLLMObsWriter class to submit spans and evals #12966, we used our own
`_get_connection()` to return HTTP/HTTPS connections. However we forgot
to include UDSHTTP connection (for the unix socket case), which means we
broke UDS support until now.

### Why was this a problem in the first place? 
We used our own `_get_connection()` in #12966 because of an issue where
creating the shared HTTPConnection helper class was leading to MRO
superclass constructor issues in our tests. At the time we thought this
was due to the shared HTTPConnection helper class having multiple
superclasses and an issue with Python 3.10 in general, but this turns
out to be due to vcrpy mocking HTTPConnection entirely and only being an
issue in tests that rely on vcrpy. This PR makes some changes to avoid
using vcrpy when not necessary, and making better assertions to ensure
that spans are being sent (not necessary in most tests to have them be
accepted).

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

[](https://datadoghq.atlassian.net/browse/MLOB-2725)

---------

Co-authored-by: IAL32 <me@adct.it>