chore(iast): improve detection of ddtrace frames (#13344)
Use more robust method to filter out frames from `ddtrace` module. The
previous version would not work in some tests if the relevant frame was
in site-packages within a virtual env named `ddtrace`, which happens in
tests for #13256. This would generally not affect customers, but
misbehaves in our test suite.