refactor(profiler): clear frozen map instead of using invalid iterator (#13353)
The cwisstable docs say conflicting things about what happens to an
iterator after an erase_at call. The map_api.h example from the repo
says the iterator is invalid. The internal CWISS_RawTable_erase_at
function, which is source of truth, says the iterator is invalid after
the call. Fix memalloc_heap_map_destructive_copy to just clear after
copying over the entries rather than erasing the iterator.

Note that this is currently dead code. We take the heap profiler lock
prior to freezing the heap profiler and emptying it out. So we can't add
an entry to the frozen array because we'll fail to acquire the lock
needed to do so. This means the map is empty and we won't see the
invalidated iterator. However, this code will run once we rework the
locking. I saw this code crash when pulling in the heap map code into my
work-in-progress branch where it actually runs. So we don't technically
_need_ to fix this right now. However, I'd like the code on main to be
right even if it's not running yet :)