fix(tracing): handle unicode when truncating long span attributes (#13475)
PR https://github.com/DataDog/dd-trace-py/pull/13270 introduced
truncation of long span attributes. However, the truncation code works
at the character level (e.g., uses `len(text)` to count the string
length), but `msgpack_pack_unicode()` expects a size in bytes as an
argument. For a string with non-ASCII characters, the character length
can be less than the byte length, and so in some cases the string would
not be truncated (because the number of characters would be below the
limit), but `msgpack_pack_unicode()` would fail (because the number of
bytes would be above the limit).

This PR changes the call to `msgpack_pack_unicode()` to use the old
limit of `ITEM_LIMIT` (2**32 - 1).
 
## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)