feat(ci_visibility): print report links at the end of the test session (#13545)
This introduces report links to the pytest plugin. At the end of a test
session, ddtrace shows links to the Datadog Test Optimization pages with
the test results for the current commit and for the current CI
job(provided that the CI environment variables with the current job and
pipeline ID are available).

This is a feature that we already have [in
`datadog-ci`](https://github.com/DataDog/datadog-ci/blob/v3.7.0/src/commands/junit/renderer.ts#L50)
for JUnit XML upload , and internally in the dogweb CI. This PR makes it
available in dd-trace-py.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)