fix: wrapping context __exit__ leak
We fix a leak in the universal wrapping context caused to bound __exit__
method leaking in CPython 3.9 and 3.10. This was due to a reference to
the bound method __exit__ of the universal wrapping context not getting
popped from the stack before every return statement. Every call to a
wrapped function would then create one such bounded method object that
would not get GC'd. No other versions of CPython are affected because
the bytecode and/or approach are different.