chore(ci): fix test spans being incorrectly marked as failed (#13574)
The repo `tests/conftest.py` defines a `pytest_runtest_protocol` which
was overriding the `pytest_runtest_protocol` hook from the Test
Optimization pytest plugin, with a number of consequences:

- Tests marked with `skip`, `skipif` or `subprocess` markers would not
run the Test Optimization hook, causing their spans not to be finished
(which in turned caused their module/suite/session not to be finished)
and left in a `failed` stated.
- The way the conftest.py hook invoked the builtin
`pytest_runtest_protocol` directly with `nextitem=None` forced some
session fixtures to be torn down and set up again in the next test,
which masked a problem with a database fixture in the Django tests.

This PR:
- Changes the conftest.py hook to return `None` instead of calling the
builtin `pytest_runtest_protocol` directly. This causes the next hook
(Test Optimization) to run instead of the builtin one.
- The Django test that uses a database in a subprocess was changed to
not drop the database at the end of the subprocess, so it will still be
available to the main process.

Future work:
- The tests with the `subprocess` marker are still being incorrectly
sent as failed to Test Optimization (the pytest session passes
normally). This will be fixed in a subsequent PR.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)