perf(rc): improve performance of subprocess service sharing
This change will only write service names to the shared file queue only if
we have not shared that service name yet.

This improvement avoids locking and writing to disk every time we call Pin.onto
which can happen a lot when database connections are cycled/recreated frequently.

It also helps ensure that the same process doesn't continually write the same service
name to the same file over and over again.