fix: iteration block hook injection
We fix an issue with the instrumentation of the beginning of iteration
blocks (e.g. for loops) whereby instrumenting the end might cause
unexpected behaviour when the modified function is invoked. This is due
to the bytecode VM not liking any bytecode interposed before the
instruction that ends the iteration block (e.g. `END_FOR`). This also
solves the problem of double-calls of the instrumented hook when hitting
the beginning of such blocks.