feat(tracer): deprecate escaped and timestamp attributes in record_exception API (#13798)
# Motivation

`record_exception` call was added few releases ago because some
customers were asking to be able to send manually errors. However, it
was mainly based on the OpenTelemetry implementation. The OTEL sdk are
inconsistent between tracers and some implementations are using for
example `escaped` attribute which does not make sense at all for us.
An
[RFC](https://docs.google.com/document/d/1-7CsG8GZVnNU198FfzavCNMI6IF0cGzFfS8cIz6Pu1g/edit?usp=sharing)
was written to ease the implementation across tracer. This PR is there
to make the python implementation consistent with the RFC.

# Changes
- Deprecates `escaped` and `timestamp` attributes
- Add a validation steps for attributes. If an attribute has not one of
the typed defined in this
[RFC](https://docs.google.com/document/d/1cVod_VI7Yruq8U9dfMRFJd7npDu-uBpste2IB04GyaQ/edit?tab=t.0#heading=h.z8rsw69ccced),
the span event will be dropped. Without span validation, it would be
"shadow" dropped which could make the user think the feature is not
working.
- Improves `record_exception` doc.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)