fix(llmobs): fix span linking for langgraph nodes (#13423)
Fix for `langgraph>=0.3.22`. This includes expanding our testing range
for `langgraph`, and it was determined we support as low as
`langgraph==0.2.23`.

Most LOC are from riot requirements fixes.

Below is some additional context for how we do span linking/fix span
linking for LangGraph. There are two other small fixes:

1. Subgraph detection was previously done by checking node names
(`"LangGraph"` equality). Since (sub)graphs can be names, this is not
viable, and instead we opt for an instance check.
2. For using the `Command`
([ref](https://blog.langchain.dev/command-a-new-tool-for-multi-agent-architectures-in-langgraph/))
and `goto` features in LangGraph will throw a `ParentCommand` exception
to communicate to the parent graph. This is not a true error, so any
instances of this error should not be tagged as such.

MLOB-2774

- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

- [x] Reviewer has checked that all the criteria below are met
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

(cherry picked from commit 25fb075)

Co-Authored-By: Sam Brenner <106700075+sabrenner@users.noreply.github.com>