feat(appsec): add metric for unsupported lambda event types (#13855)
## Motivation

Avoid billing when Appsec is enabled for unsupported lambda events. To
keep track of executions with unsupported events, we add a span metric.

## Changes

- Selectively skip processing the span based on the event

To make the information available, I used the same pattern as the asm
context initialization by storing temporary information inside the
`ExecutionContext`.

The only difference is that in the case of lambda we only have a single
global `ExecutionContext` so we have to clean it up.

## Notes

This PR relies on:
https://github.com/DataDog/datadog-lambda-python/pull/627



## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)